+++ packages/elements/src/components/ui/buttons/FlatButton.tsx
@@ -13,7 +13,7 @@ export const FlatButton: React.FC<FlatButtonProps> = ({
   ...props
 }) => (
   <PrimaryButton
-    className={cx(className, styles.flatButton, inverted && styles.inverted)}
+    className={cx(styles.flatButton, inverted && styles.inverted, className)}
     {...props}
   />
 );