+++ packages/panels/src/components/nav-bar/NavBarHeading.tsx
@@ -1,0 +1,8 @@
+import * as React from "react";
+import { Heading, HeadingProps } from "@stenajs-webui/core";
+
+interface NavBarHeadingProps extends HeadingProps {}
+
+export const NavBarHeading: React.FC<NavBarHeadingProps> = (headingProps) => {
+  return <Heading whiteSpace={"nowrap"} color={"#fff"} {...headingProps} />;
+};