+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 5.1.0
+
+### Icons
+
+- Add stories for custom icons.
+
+### NavBar
+
+- Adjusted some alignments between side menu button, left content and buttons.
+- Centered content is now always centered.
+- New component `NavBarSearchField` which is a custom style input field.
+- New component `NavBarHeading` which is a heading text for nav bar.
+
 ## 5.0.0
 
 ### FlatButton